refactor to remove redundant code
